Move VCS events APIs to Modules
